Add tt restart and status commands
This patch adds two new tt command to the command pallete: restart and
status.

The `restart` command is still problematic since it doesn't support
interactive flag making it impossible to restart the cluster started by
the `start` command with preserving the output.

Also it introduces minor checks ensuring the `tt` is installed.