Fix missing init extension command
This patch fixes a bug appeared in f39a34bb ('Change package name to
`tarantool`') where the `init` command was not renamed with
`tarantool` prefix and has been left with `tarantool-vscode` making the
command broken.

This patch renames the command declaration to `tarantool.init-vs` and
fixes the problem.

No changelog is added since the change hasn't been released yet.