Fix wrong artifact path in CI publish job
This patch fixes a wrong path in publish CI job. Turned out the
artifacts are downloaded to `./<artifact-name>/<artifact>` and not in
`./<artifact>` as it's been expected. Thus, the artifact has been
downloaded in the unexpected directory.

Let's rename the artifact name to `tarantool-vscode` for the job to
seek in the same directory name as the `.vsix` name.